Add necessary packages for EF and create DB context file

+++ India.Domain/Models/City.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.ComponentModel.DataAnnotations.Schema;
 using System.Linq;
-using System.Numerics;
 using System.Text;
 using System.Threading.Tasks;
 
@@ -11,7 +11,7 @@ namespace India.Domain.Models
     {
         public Guid Id { get; set; }
         public string Name { get; set; }
-        public BigInteger Population { get; set; }
+        public long Population { get; set; }
         public Guid StateId { get; set; }
 
         //navigation properties
